Hook up DB with our app and create User model
We need to hook up the db with our app.
For that, needed to do following:
1. We initialize db extension when the app is created.
2. Have the config settings define the URI for the database so
SQLAlchemy can find our db.
3. Define a User model with the columns that we need. Note that we
just need to specify things as object, sqlalchemy/psycopg will deal
with postgres.
4. We need a way to interact with the db other than through the app,
and for that we created a CLI for the application. The CLI provides
option for interacting with db (to reset initialize etc.).
5. When the DB is reset a seed admin user is automatically created.