Added JWT based authorization to API end points.
Added jwt token requirement for update & delete end points.

Initially, we wanted to test this with some command line tests, but
that was becoming a bit cumbersome, so we used postman to do the
testing.

On the postman side, for the requests to work, cookies (csrf and access
token) needs to be saved in the environment so that subsequent calls
can reuse it.